Reorganize errors logic
