+++ src/logic/index.js
@@ -2,12 +2,14 @@ import linkGithubIdentity from './linkGithubIdentity';
 import logout from './logout';
 import startAccountMigration from './startAccountMigration';
 import unlinkGithubIdentity from './unlinkGithubIdentity';
-import errors from './errors';
+import validateCurrentProject from './validateCurrentProject';
+import validateProjectOnChange from './validateProjectOnChange';
 
 export default [
   linkGithubIdentity,
   logout,
   startAccountMigration,
   unlinkGithubIdentity,
-  ...errors,
+  validateCurrentProject,
+  validateProjectOnChange,
 ];
